api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

sqlalchemy==2.0.36
psycopg[binary]==3.2.3

google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

itsdangerous==2.2.0
requests==2.32.3
email-validator==2.2.0
openai
requests



